organized client-side in src/client and server-side in src/server
