+++ .babelrc
@@ -8,7 +8,7 @@
         "react-hot-loader/babel",
         "inline-react-svg",
         ["module-resolver", {
-            "root": ["./src"]
+            "root": ["./src/client"]
         }
         ]
     ]
