+++ src/client/.eslintrc.json
@@ -11,6 +11,7 @@
     "rules": {
       "react/jsx-filename-extension": 0,
       "import/no-extraneous-dependencies": 0,
+      "no-underscore-dangle": 0,
       "global-require": 0
     },
     "settings": {
