+++ src/client/assets/images/favicon.png
