+++ src/client/index.jsx
@@ -1,4 +1,3 @@
-import 'babel-polyfill';
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
@@ -7,9 +6,9 @@ import { AppContainer } from 'react-hot-loader';
 import AppRouter from 'containers/app-router';
 import configureStore from './redux/store/store';
 
-const preloadedState = window.__PRELOADED_STATE__; //eslint-disable-line
+const preloadedState = window.__PRELOADED_STATE__;
 
-delete window.__PRELOADED_STATE__; //eslint-disable-line
+delete window.__PRELOADED_STATE__;
 
 const store = configureStore(preloadedState);
