+++ src/server/.eslintrc.json
